+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/services/RootsManagerService.kt
@@ -1,0 +1,67 @@
+package org.plan.research.minimization.plugin.services
+
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.roots.ProjectRootManager
+import com.intellij.openapi.vfs.VfsUtil
+import com.intellij.openapi.vfs.VirtualFile
+
+@Service(Service.Level.PROJECT)
+class RootsManagerService(private val rootProject: Project) {
+    private fun propagateAndMergeRoots(
+        contentRoots: List<VirtualFile>,
+        srcRoots: List<VirtualFile>,
+        sourceRoots: List<VirtualFile>,
+    ): List<VirtualFile> {
+        // propagate src roots (replace them with their children) if it contains any content root
+        val propagationStatus = HashMap<VirtualFile, PropagationStatus>()
+        for (contentRoot in contentRoots) {
+            propagationStatus[contentRoot] = PropagationStatus.IS_CONTENT_ROOT
+            var parent: VirtualFile? = contentRoot.parent
+            while (parent != null) {
+                propagationStatus.putIfAbsent(parent, PropagationStatus.NEED_TO_PROPAGATE) ?: break
+                parent = parent.parent
+            }
+        }
+
+        val queue = ArrayDeque<VirtualFile>()
+        queue.addAll(srcRoots)
+
+        val roots = mutableListOf<VirtualFile>()
+        while (queue.isNotEmpty()) {
+            val root = queue.removeFirst()
+            val status = propagationStatus[root]
+            propagationStatus[root] = PropagationStatus.ALREADY_PROPAGATED_OR_ADDED
+            when (status) {
+                PropagationStatus.NEED_TO_PROPAGATE -> queue.addAll(root.children)
+                PropagationStatus.IS_CONTENT_ROOT, PropagationStatus.ALREADY_PROPAGATED_OR_ADDED -> {}
+                else -> roots.add(root)
+            }
+        }
+
+        // delete all sourceRoots that are in any of the already added roots
+        val sourceRootsToAdd = sourceRoots.filter { sourceRoot ->
+            roots.none { src -> VfsUtil.isAncestor(src, sourceRoot, false) }
+        }
+        roots.addAll(sourceRootsToAdd)
+
+        return roots
+    }
+
+    fun findPossibleRoots(): List<VirtualFile> {
+        val rootManager = ProjectRootManager.getInstance(rootProject)
+
+        val sourceRoots = rootManager.contentSourceRoots.toList()
+        val contentRoots = rootManager.contentRoots.toList()
+        val srcRoots = contentRoots.mapNotNull { it.findChild("src") }
+
+        return propagateAndMergeRoots(contentRoots, srcRoots, sourceRoots)
+    }
+
+    private enum class PropagationStatus {
+        ALREADY_PROPAGATED_OR_ADDED,
+        IS_CONTENT_ROOT,
+        NEED_TO_PROPAGATE,
+        ;
+    }
+}
